move ASWA to weight averaging.py
