+++ dicee/weight_averaging.py
@@ -10,6 +10,141 @@ from .abstracts import AbstractCallback
 from dicee.models.ensemble import EnsembleKGE
 from .eval_static_funcs import evaluate_ensemble_link_prediction_performance
 
+class ASWA(AbstractCallback):
+    """ Adaptive stochastic weight averaging
+        ASWE keeps track of the validation performance and update s the ensemble model accordingly.
+        """
+
+    def __init__(self, num_epochs, path):
+        super().__init__()
+        self.path=path
+        self.num_epochs=num_epochs
+        self.initial_eval_setting = None
+        self.epoch_count=0
+        self.alphas = []
+        self.val_aswa = -1
+
+    @rank_zero_only
+    def on_fit_end(self, trainer, model):
+        # super().on_fit_end(trainer, model)
+        if self.initial_eval_setting:
+            # ADD this info back
+            trainer.evaluator.args.eval_model = self.initial_eval_setting
+        
+        param_ensemble = torch.load(f"{self.path}/aswa.pt", torch.device("cpu"))
+        model.load_state_dict(param_ensemble)
+
+    @staticmethod
+    def compute_mrr(trainer, model) -> float:
+        # (2) Enable eval mode.
+        eval_model = copy.deepcopy(model)
+        eval_model.eval()
+        # (3) MRR performance on the validation data of running model.
+        eval_model.to("cpu")
+        last_val_mrr_running_model = trainer.evaluator.eval(dataset=trainer.dataset,
+                                                            trained_model=eval_model,
+                                                            form_of_labelling=trainer.form_of_labelling,
+                                                            during_training=True)["Val"]["MRR"]
+        del eval_model
+        return last_val_mrr_running_model
+
+    def get_aswa_state_dict(self, model):
+        # (2) Question: Soft update or Rejection?!
+        if isinstance(model,EnsembleKGE):
+            ensemble_state_dict = torch.load(f"{self.path}/aswa.pt")
+        else:
+            ensemble_state_dict = torch.load(f"{self.path}/aswa.pt",  torch.device(next(model.parameters()).device))
+        # Perform provision parameter update.
+        with torch.no_grad():
+            for k, parameters in model.state_dict().items():
+                if parameters.dtype == torch.float:
+                    ensemble_state_dict[k] = (ensemble_state_dict[k] * sum(self.alphas) + parameters) / (1 + sum(self.alphas))
+        return ensemble_state_dict
+
+    def decide(self, running_model_state_dict, ensemble_state_dict, val_running_model, mrr_updated_ensemble_model):
+        """
+        Perform Hard Update, software or rejection
+
+        Parameters
+        ----------
+        running_model_state_dict
+        ensemble_state_dict
+        val_running_model
+        mrr_updated_ensemble_model
+
+        Returns
+        -------
+
+        """
+        # (1) HARD UPDATE:
+        # If the validation performance of the running model is greater than
+        # the validation performance of updated ASWA and
+        # the validation performance of ASWA
+        if val_running_model > mrr_updated_ensemble_model and val_running_model > self.val_aswa:
+            """Hard Update """
+            # (1.1) Save the running model as ASWA
+            torch.save(running_model_state_dict, f=f"{self.path}/aswa.pt")
+            # (2.1) Resect alphas/ensemble weights
+            self.alphas.clear()
+            # (2.2) Store the validation performance of ASWA
+            self.val_aswa = val_running_model
+            return True
+
+        # (2) SOFT UPDATE:
+        # If the validation performance of the running model is less  than
+        # the validation performance of updated ASWA
+        if mrr_updated_ensemble_model > self.val_aswa:
+            """Soft update"""
+            self.val_aswa = mrr_updated_ensemble_model
+            torch.save(ensemble_state_dict, f=f"{self.path}/aswa.pt")
+            self.alphas.append(1.0)
+            return True
+        # (3) Rejection:
+        if self.val_aswa > mrr_updated_ensemble_model:
+            """ Ignore """
+            self.alphas.append(0)
+            return True
+
+    @rank_zero_only
+    def on_train_epoch_end(self, trainer, model):
+        # if (trainer.global_rank == trainer.local_rank == 0) is False:
+        #     return None
+        if isinstance(model, OptimizedModule):
+            model = model._orig_mod
+        # (1) Increment epoch counter
+        self.epoch_count += 1
+        # (2) Save the given eval setting if it is not saved.
+        if self.initial_eval_setting is None:
+            self.initial_eval_setting = trainer.evaluator.args.eval_model
+            trainer.evaluator.args.eval_model = "val"
+        # (3) Compute MRR of the running model.
+        val_running_model = self.compute_mrr(trainer, model)
+
+        # (4) Initialize ASWA if it is not initialized.
+        if self.val_aswa == -1:
+            torch.save(model.state_dict(), f=f"{self.path}/aswa.pt")
+            self.alphas.append(1.0)
+            self.val_aswa = val_running_model
+            return True
+        else:
+            # (5) Load ASWA ensemble parameters.
+            ensemble_state_dict = self.get_aswa_state_dict(model)
+            # (6) Initialize ASWA ensemble with (5).
+            ensemble = copy.deepcopy(model)
+            ensemble.load_state_dict(ensemble_state_dict)
+            ensemble.to("cpu")
+
+                
+            # (7) Evaluate (6) on the validation data, i.e., perform the lookahead operation.
+            mrr_updated_ensemble_model = trainer.evaluator.eval(dataset=trainer.dataset,
+                                                                trained_model=ensemble,
+                                                                form_of_labelling=trainer.form_of_labelling,
+                                                                during_training=True)["Val"]["MRR"]
+            # print(f"| MRR Running {val_running_model:.4f} | MRR ASWA: {self.val_aswa:.4f} |ASWA|:{sum(self.alphas)}")
+            # (8) Decide whether ASWA should be updated via the current running model.
+            del ensemble
+            self.decide(model.state_dict(), ensemble_state_dict, val_running_model, mrr_updated_ensemble_model)
+
 class SWA(AbstractCallback):
     """Stochastic Weight Averaging callback.
